LayerToolのロジックをカスタムフックに切り出すリファクタ
